external signal works!

+++ cilantro_ee/messages/capnp/signals.capnp
@@ -7,7 +7,7 @@ struct Signal {
 
 struct ExternalSignal {
     id @0 :UInt32;
-    timestamp @1: Float32;
+    timestamp @1: UInt32;
     sender @2: Data;
     signature @3: Data;
 }
